+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/shared/viewport-popup/viewport-popup.component.ts
@@ -1,5 +1,5 @@
 import type { OnInit } from '@angular/core';
-import { Component, EventEmitter, Output } from '@angular/core';
+import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 import type { UUID, Viewport } from 'digital-fuesim-manv-shared';
 import type { Observable } from 'rxjs';
@@ -7,6 +7,7 @@ import { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
 import { createSelectViewport } from 'src/app/state/application/selectors/exercise.selectors';
 import { selectCurrentRole } from 'src/app/state/application/selectors/shared.selectors';
+import { PopupService } from '../../utility/popup.service';
 
 @Component({
     selector: 'app-viewport-popup',
@@ -17,14 +18,13 @@ export class ViewportPopupComponent implements OnInit {
     // These properties are only set after OnInit
     public viewportId!: UUID;
 
-    @Output() readonly closePopup = new EventEmitter<void>();
-
     public viewport$?: Observable<Viewport>;
     public readonly currentRole$ = this.store.select(selectCurrentRole);
 
     constructor(
         private readonly store: Store<AppState>,
-        private readonly exerciseService: ExerciseService
+        private readonly exerciseService: ExerciseService,
+        private readonly popupService: PopupService
     ) {}
 
     ngOnInit() {
@@ -39,5 +39,9 @@ export class ViewportPopupComponent implements OnInit {
             viewportId: this.viewportId,
             newName,
         });
+    }
+
+    public closePopup() {
+        this.popupService.closePopup();
     }
 }