+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/simulated-region-overview.module.ts
@@ -46,6 +46,8 @@ import { RadigoramCardContentResourceRequestComponent } from './radiogram-list/r
 import { SimulatedRegionOverviewBehaviorAutomaticallyDistributeVehiclesComponent } from './tabs/behavior-tab/behaviors/automatically-distribute-vehicles/simulated-region-overview-behavior-automatically-distribute-vehicles.component';
 import { RequestVehiclesComponent } from './tabs/behavior-tab/behaviors/request-vehicles/simulated-region-overview-behavior-request-vehicles.component';
 import { SimulatedRegionOverviewPatientInteractionBarComponent } from './tabs/patients-tab/simulated-region-overview-patient-interaction-bar/simulated-region-overview-patient-interaction-bar.component';
+import { SimulatedRegionOverviewVehiclesTabComponent } from './tabs/vehicles-tab/simulated-region-overview-vehicles-tab.component';
+import { SimulatedRegionOverviewPatientsTableComponent } from './patients-table/simulated-region-overview-patients-table.component';
 
 @NgModule({
     declarations: [
@@ -83,6 +85,8 @@ import { SimulatedRegionOverviewPatientInteractionBarComponent } from './tabs/pa
         SimulatedRegionOverviewBehaviorAutomaticallyDistributeVehiclesComponent,
         RequestVehiclesComponent,
         SimulatedRegionOverviewPatientInteractionBarComponent,
+        SimulatedRegionOverviewVehiclesTabComponent,
+        SimulatedRegionOverviewPatientsTableComponent,
     ],
     imports: [
         FormsModule,