+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/tabs/behavior-tab/behaviors/request-vehicles/simulated-region-overview-behavior-request-vehicles.component.ts
@@ -58,7 +58,7 @@ export class RequestVehiclesComponent implements OnChanges {
                     const options = Object.entries(simulatedRegions)
                         .map(([id, simulatedRegion]) => [
                             id,
-                            simulatedRegion.name,
+                            `[Simuliert] ${simulatedRegion.name}`,
                         ])
                         .filter(([id, _name]) => id !== this.simulatedRegionId)
                         .sort(([id1, name1], [id2, name2]) =>