+++ shared/src/simulation/behaviors/treat-patients.ts
@@ -158,8 +158,12 @@ export const treatPatientsBehavior: SimulationBehavior<TreatPatientsBehaviorStat
                     break;
                 case 'treatmentProgressChangedEvent':
                 case 'materialAvailableEvent':
+                case 'materialRemovedEvent':
                 case 'newPatientEvent':
+                case 'patientRemovedEvent':
                 case 'personnelAvailableEvent':
+                case 'personnelRemovedEvent':
+                case 'leaderChangedEvent':
                 case 'treatmentsTimerEvent': {
                     if (event.type === 'treatmentProgressChangedEvent') {
                         behaviorState.treatmentProgress = event.newProgress;