+++ shared/src/store/action-reducers/patient.ts
@@ -11,6 +11,7 @@ import {
     currentSimulatedRegionIdOf,
     currentCoordinatesOf,
     isInSimulatedRegion,
+    currentSimulatedRegionOf,
 } from '../../models/utils';
 import {
     changePosition,
@@ -33,10 +34,23 @@ import { updateTreatments } from './utils/calculate-treatments';
 import { getElement } from './utils/get-element';
 import { removeElementPosition } from './utils/spatial-elements';
 
+/**
+ * Performs all necessary actions to remove a patient from the state.
+ * This includes deleting all treatments, removing it from the spatial tree and sending a {@link PatientRemovedEvent} if the patient is in a simulated region.
+ * @param patientId The ID of the patient to be deleted
+ */
 export function deletePatient(
     draftState: Mutable<ExerciseState>,
     patientId: UUID
 ) {
+    const patient = getElement(draftState, 'patient', patientId);
+    if (isInSimulatedRegion(patient)) {
+        const simulatedRegion = currentSimulatedRegionOf(draftState, patient);
+        sendSimulationEvent(
+            simulatedRegion,
+            PatientRemovedEvent.create(patientId)
+        );
+    }
     removeElementPosition(draftState, 'patient', patientId);
     delete draftState.patients[patientId];
 }